feat: 🔥💩 Fix Docker build workflow to handle '**' branch and tag with '*', causing unnecessary re-repos...
I spent hours trying to figure out why my Docker build script was going crazy, only to realize that it was because of this stupid '**' branch in our workflow. But no, the developers had to go and mess with the git hooks too. Fixing this issue took me less than 10 minutes. Now we can all breathe a sigh of relief.